Dockerized Application - Timezone logic updated
The application can be started now via docker containers.

```bash
docker-compose build && docker-compose up
```

Will start the app and be accessible via `http://localhost:4567`.

`.env` is required and it must contain the necessary values as described
in the README.md

The time used to calculate the Sunset/Sunrise values have been updated
to include the time offset values that are returned from the Weather
API. This meant that those values (Sunset/Sunrise) are assumed to be UTC
values and the "timezone" value is added against it.