Mobile CSS (+Bootstrap Update) + Midnight Fix
Font sizes were set to only display a fixed pt size that caused issues
when trying to view the App on smaller screen sizes (like mobile
devices). Updating to Bootstrap 4.x and converting the font-sizes for
the content to use `vw` calculations - we can support multiple
resolutions.

Midnight fix is the issue where if the clock was 0 hour (meaning
midnight), it would display 0:00. The fix now is to ensure that it is
set to 12 so that midnight is displayed as "12:00".